test: centralize project export setup
